requirements.txt
fastapi
pydantic
uvicorn
PyJWT
requests
umsgpack
pycryptodome
cryptography
httpx